fixed conversion warnings, and hopefully the file include paths

+++ tmpl8-2026/gameObject.cpp
@@ -48,10 +48,10 @@ void GameObject::DrawOrigin()
 	vec2 screenPos = pos - offset;
 
 	Central::surface->Box(
-		round(screenPos.x - 2), // Rounding keeps box size consistent - truncation causes jitter
-		round(screenPos.y - 2), 
-		round(screenPos.x + 2), 
-		round(screenPos.y + 2),
+		(int)round(screenPos.x - 2), // Rounding keeps box size consistent - truncation causes jitter
+		(int)round(screenPos.y - 2),
+		(int)round(screenPos.x + 2),
+		(int)round(screenPos.y + 2),
 		0xFF0000);
 }
